formal: close the audit's follow-up round (R-1, R-2, R-3, and the B-4 generalization)
R-1 (the report is right; our first C-3 round overstated): linearization_vesta.json
regenerates byte-identical and still had gate_combined.endoMul = varBaseMul = "0", so
check_linearization's per-gate checks for exactly the two gates V-1 concerned were
0 = 0, with coverage resting only on whole-proof acceptance.

* linearization_dump now emits a SECOND fixture from the emul circuit,
  linearization_vesta_emul.json — live gate terms varBaseMul and endoMul. The emul
  circuit moved to lib.rs so both dumpers share it. The historical fixture stays
  byte-identical: the prover must continue the SAME rng the circuit drew from, which
  the refactor threads through rather than reseeding.
* check_linearization runs both fixtures and each declares which gates it exercises;
  a zero target for a declared gate now FAILS the driver, and the report annotates
  zero targets '(0)' so a vacuous check reads as vacuous. Every gate is live in at
  least one fixture.
* NC-2: under the pre-fix EndoMul the mixed fixture passes unchanged while the emul
  fixture reports 'endoMul: ✗' beside 'varBaseMul: ✓' — localized to the gate, which
  is what R-1 asked for.

R-2: docs/negative-controls.md records each discrimination check — the mutation, the
observed failure, replayable — for V-1 end-to-end (NC-1), R-1 localization (NC-2,
including that a sign-only perturbation fails to compile because holds_iff cross-checks
the list), V-2's trace shape (NC-3), and the new exhibit guards (NC-4).

R-3: both §2.1 observations fixed rather than accepted. The pull_request trigger no
longer filters on base branch, so a stacked PR is gated like any other; setup-lean
builds from the committed lake-manifest.json, with the re-resolve surviving only as a
fallback that warns the run is not on the committed pin.

The addendum's B-4 generalization — under a dead=0 gate any exhibit absent from
roots.txt is indistinguishable from dead code — is acted on: both locked-target gates
now pin the EXISTENCE of the whole exhibit set (20 bulletproof-pcs, 6 kimchi). Rooting
alone only protects an exhibit while it stays rooted; a sweep removing root and
declaration together was still silently green. Deleting one now fails a lock.
