+++ formal/kimchi/fixtures/linearization_vesta_emul.json
@@ -1,0 +1,221 @@
+{
+  "alpha": "22730214477862889077433839509717939897171364934800124428623675293366817551295",
+  "beta": "296558289474281640746133544226371517309",
+  "coefficients": [
+    [
+      "0",
+      "0"
+    ],
+    [
+      "0",
+      "0"
+    ],
+    [
+      "0",
+      "0"
+    ],
+    [
+      "0",
+      "0"
+    ],
+    [
+      "0",
+      "0"
+    ],
+    [
+      "0",
+      "0"
+    ],
+    [
+      "0",
+      "0"
+    ],
+    [
+      "0",
+      "0"
+    ],
+    [
+      "0",
+      "0"
+    ],
+    [
+      "0",
+      "0"
+    ],
+    [
+      "0",
+      "0"
+    ],
+    [
+      "0",
+      "0"
+    ],
+    [
+      "0",
+      "0"
+    ],
+    [
+      "0",
+      "0"
+    ],
+    [
+      "0",
+      "0"
+    ]
+  ],
+  "complete_add_selector": [
+    "0",
+    "0"
+  ],
+  "constant_term": "8191801777589774368290858758409663832738902088144923046411786010968804807120",
+  "curve": "vesta",
+  "emul_selector": [
+    "16563964795222873887529703456226991338514805751528846966466250939875436757689",
+    "9489218168424090033435947351916618966267105441939654062266177374357611311392"
+  ],
+  "endo": "20444556541222657078399132219657928148671392403212669005631716460534733845831",
+  "endomul_scalar_selector": [
+    "0",
+    "0"
+  ],
+  "ft_eval0": "2342068304258800988433301733628483469766444178417446096601543285111352511896",
+  "ft_eval1": "11109229799063182359333909712103378840967604855368514552704819461683857525756",
+  "gamma": "275718911916844720960749290730120522582",
+  "gate_combined": {
+    "completeAdd": "0",
+    "endoMul": "17285386308494160164877636600147802707915361568275938031807334840840535625104",
+    "endoScalar": "0",
+    "generic": "0",
+    "poseidon": "0",
+    "varBaseMul": "19854437778424663059305968410433838088186597001810545730559127934478236812353"
+  },
+  "generic_selector": [
+    "0",
+    "0"
+  ],
+  "index_terms": {},
+  "mul_selector": [
+    "1174422923499682973009710035668289717679237481131155608498355638519472531510",
+    "15434425429141355213895452544600012374541141132306622953953189411176951430364"
+  ],
+  "n": "16",
+  "omega": "10810933018019022645195840712901647754980539521123437677405037345586003209325",
+  "perm_scalar": "21132421002230394661739862717965412709189456722829183673903246340990901886472",
+  "poseidon_selector": [
+    "0",
+    "0"
+  ],
+  "public_evals": [
+    [
+      "0"
+    ],
+    [
+      "0"
+    ]
+  ],
+  "s": [
+    [
+      "14213706472025004769788383295830762501073683360239512852055157145621273788686",
+      "26839130656642813029364102812852540016559985066850276137118468902876204047373"
+    ],
+    [
+      "24414198687216668741036593087836465763833904073435836656210879256302070432332",
+      "325244432933680084017018405673382070718624799538006014880589813782277441357"
+    ],
+    [
+      "25201664425963568624525676688001831704230367067331808163029681935334991755657",
+      "14361130243484584297757886952231426054853623531662456132035512805249178797090"
+    ],
+    [
+      "25655530840128354557339059860716421781470911441456396738315878248612482343557",
+      "22988949955197538611387235919574762601800016044761797862890014407957977126820"
+    ],
+    [
+      "9958286210530648099242548233878072603744610086618339910986407913881634848804",
+      "1935469849350591097752503297632705736758940236204150197223194719814770660346"
+    ],
+    [
+      "3573935006827177156876758619979609326179559294251543604714689836176713669397",
+      "21082594086575158827814375467883354429853456159189367719983017635653129393995"
+    ]
+  ],
+  "shifts": [
+    "1",
+    "328286983623303317637963920346571898945724874896624808297627776768640590563",
+    "91433028157768305433241271390810941046493237899366836746431422160024463706",
+    "240213425742950025341713987028051046476975246675775993287051503548513551377",
+    "417757293700961807788464308236931191792053554682199437460107260306038610067",
+    "430348682428487492383428014506756320686619984007091686553051322507181255952",
+    "326625242707153437805405281465150497418605074624614708160829052937679007395"
+  ],
+  "sigma6_zeta": "12879517728384193602021939344141245292700975794588911166097340257252583992629",
+  "w": [
+    [
+      "4241515091590845564903044809451738041535557631136058187998346717535665316889",
+      "8495933719155878279415603697395308979649598053838116994742899779570549337160"
+    ],
+    [
+      "18699248569062101924684477949645986645798726966714749660458452885780437342456",
+      "4315654068992167008088332441416224333207438976152202080841448166006331318265"
+    ],
+    [
+      "25059375989236169392190150601306522405168013655121493686053559277542081713299",
+      "20426929717096698168607751328972233404903755988694565253695916796361464044256"
+    ],
+    [
+      "11100869495129097274063452702813993988268756052067343937015011339463501804936",
+      "11969296913531283980467280785771551236817812255976004886414339762552375763782"
+    ],
+    [
+      "13680872650429080202658007137150115409419626536976113722766040732181711618776",
+      "16242206440926356809846123172545451882912506806344415169106663349990845759124"
+    ],
+    [
+      "28619336659120635305946826582782457639369507554612549282616769630363654481582",
+      "28264998535449021835061671798357405046588756313103715247777146761284773324049"
+    ],
+    [
+      "20731555280324659302967819006657244220749407891570577559250045714113214564460",
+      "8298995919412910074300688763644590351757508135850109586681287618261513797830"
+    ],
+    [
+      "3420456788229400113110180311804656443615157191918272395502379252066057840562",
+      "1821584379975881255957393633090049512404541581992901463492309859658450012198"
+    ],
+    [
+      "25249134948458065155098510814502086332829374870845467704855500153492511416722",
+      "20660633459262663740068109962693387900023191395397093771711403176144837893947"
+    ],
+    [
+      "6525265066901007819497761601790917045084881410736299123186429555956367692749",
+      "23153459575463839683881548789937182902479514077633284238524214139095254450083"
+    ],
+    [
+      "3418972257878416807274681684073945390677084044088351570076327511884041510495",
+      "6375294471442071884223384300296881074334182442999036760705981052755929182646"
+    ],
+    [
+      "5761339192665585435332634924414554038056649642727368817716870228090167079788",
+      "6434616084812613988261338215845390774503441824905521656096259523880927625275"
+    ],
+    [
+      "1498496156012312825221547576874136561609166534977276376170140734118969232258",
+      "17312963801188083129739282082876058459839753927108845566117993231256110204444"
+    ],
+    [
+      "18154997886398620376844009976642682386512381820810470032796457362846254865656",
+      "262263470245247380461081641531910873005901124632843387207696531479347578641"
+    ],
+    [
+      "17161883858708184952242747779473852613568784496589583014173604150823331331427",
+      "12676829242973631203284803475742392945372244321855271329691668330817808884736"
+    ]
+  ],
+  "z": [
+    "19781959801711181978672573786515419392831767560350214568791630353830948031280",
+    "9535821130811108378093664245813627414049524354762724489731497214153032092815"
+  ],
+  "zeta": "14213706472025004769788383295830762501073683360239512852055157145621273788686",
+  "zk_rows": "3",
+  "zkpm_zeta": "24679386057679642859812894004751684350259950313186008729013508991306283936481"
+}
