+++ formal/scripts/fixtures.sha256
@@ -27,6 +27,7 @@ db86d6b387910fef12e24f5d2d9527a495de18830b0de6db5e5c6648460243ca  kimchi/fixture
 2ef4b1e7f802f490967cfdc817379a8e7faff1999eb414d40157f90fdde444f0  kimchi/fixtures/kimchi_proof_vesta_nc8.json
 31faac1dda037bca3d2c26556ca0247da81f9a4fcfe555758c0418190f923d67  kimchi/fixtures/kimchi_proof_vesta_pub.json
 ab577f4244a7c9820971c4ca2f1f1071a7d5bfa373e2d3d7a0a34a08a6d874cc  kimchi/fixtures/linearization_vesta.json
+ea8a77ba3c209c710b75d0a19965109a27307899eb6f18ff87a3ca4193594080  kimchi/fixtures/linearization_vesta_emul.json
 2a879a61c54363550a8fcd640911546180f99c9982f1c6938893ea6ced7d185e  kimchi/fixtures/perm_vesta.json
 d3b09ea5aa0ddbcecfe5fe12fc6dba20c0b6f2b446bdd2980c8acd9d74bdc2d5  poseidon/fixtures/fq_sponge_pallas_vectors.json
 bab1beee0c88172c018ac68df8d98f5b6e29e19a4a2fd861e8318c4405445f86  poseidon/fixtures/fq_sponge_vectors.json
